+++ CHANGELOG.md
@@ -65,6 +65,7 @@ These classes will run as threads in the next releases:
 
 - [s6-rc]: Delete matter.js storage locks before starting the s6-rc and s6-rc-legacy docker images.
 - [test-utils]: Add matter methods in jest-utils and vites-utils packages.
+- [soilMeasurement]: Add createDefaultSoilMeasurementClusterServer() endpoint helper.
 
 ### Changed